refactor: move carbon estimation table to components
